+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 ## [Unreleased]
 
 ### Added
+1. Added `get_writer` to `ms_deisotope.output` which will guess which writer (and compression) to use from a file path
+   or name of a file-like object.
 
 ### Changed